microservices-basketball/players-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
pymongo==4.6.1
mongoengine==0.28.2
python-dotenv==1.0.0
marshmallow==3.20.2
marshmallow-mongoengine==0.30.0
requests==2.31.0
gunicorn==21.2.0
werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
black==23.12.1
flake8==7.0.0
